pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.1",
    "sqlalchemy>=2.0.35",
    "werkzeug>=3.0.4",
    "wtforms>=3.1.2",
    "scikit-learn>=1.5.2",
    "pandas>=2.2.3",
    "numpy>=2.1.2",
]
